Patch the ganeti-os-interface man page
This will hopefully go upstream too, but it's not worth waiting for the next
release for us to upload so we temporarily add a debian patch system. Since we
don't plan to manage lots of patches the simple cdbs patch system has been
chosen. We can convert to quilt or dpatch later if we think we need to.