Fix the startup-with-1.2-config (Closes: #539303)
Before this patch, we abort if we don't find a new 2.0 ssconf key. But
this is wrong, as it breaks noded startup at cluster init (when no
ssconf keys exist at all). This patch changes so that we abort if a
1.2-specific ssconf key if found (reverse test).

The patch also changes the exit code to show failure, so that errors are
more easily detected.