postinst: remove stale PID and socket files
Due to upstream issues #581 and #582, upgrades from previous versions will fail
to start the Haskell daemons because of wrong PID and socket file ownership. As
a workaround, we remove the stale files in postinst. No checks against the
daemons running are done, the daemons are expected to have been stopped during
prerm.