postinst: fix job queue archive permissions
ensure_dirs by default does not recursively change the job queue's archive
directory permissions. As a result, when master runs as a non-root user, it
fails to archive jobs during a watcher run.

We fix this by calling ensure_dirs -f late in postinst, only if it seems we
have to (as it may take a while).