Turn ganeti-htools into recommends with strict versioning
htools is a much more integral part of Ganeti than it used to be and user
experience is incomplete without it.

Furthermore, since compatibility between different versions of Ganeti and
htools is no longer guaranteed, we also break on older and newer htools
versions to force the same version of htools as ganeti.