Prompt to abort removal of versioned packages
Use debconf to prompt the user to abort removal of versioned packages
that are still in use. This scenario can come up after upgrading to a
new minor version and before running `gnt-cluster upgrade'. In this
case, `apt-get autoremove' will try to remove the old packages, which
however are still in use.